+++ persistence/sql/migrations/sql/20220420102701000000_identity_metadata.mysql.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE identities ADD metadata_public JSON NULL;
+ALTER TABLE identities ADD metadata_admin JSON NULL;