+++ selfservice/strategy/oidc/provider_private_net_test.go
@@ -3,6 +3,7 @@ package oidc_test
 import (
 	"context"
 	"fmt"
+	"net/url"
 	"testing"
 	"time"
 
@@ -77,7 +78,7 @@ func TestProviderPrivateIP(t *testing.T) {
 			p := tc.p(tc.c)
 			_, err := p.Claims(context.Background(), (&oauth2.Token{RefreshToken: "foo", Expiry: time.Now().Add(-time.Hour)}).WithExtra(map[string]interface{}{
 				"id_token": tc.id,
-			}))
+			}), url.Values{})
 			require.Error(t, err)
 			assert.Contains(t, fmt.Sprintf("%+v", err), tc.e)
 		})